backend/api-gateway/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-socketio>=5.11.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
